refactor RAIInsights into RAIInsightsBase class for basic functionality (#1284)
